Fix (partly) the focus stealing issue
Giving focus to the TCP after each mouse click seems to work fine.
However, after scrolling it means that no modifier keys are given
to the script. INstead, the TCP seems to intercept all modifier
keys itself. So this works except for when NO modifier keys are
down when rolling the mouse wheel. Then scroll messages are not
interpreted as expected.