Fix mouse wheeling focus issue
A script receives mouse wheel commands when wheeled over while not
having focus, but does not receive modifier keys when not focused.
When wheeled over, the script should behave differently depending
on the modifier key(s) held down when wheeling. So, when the not
having focus and being wheeled over, the script grabs the focus
and delays by one defer cycle the action on that wheel command. A
timer is set to count down during inactive time to give the focus
back to teh TCP after the count down. This is to let key strokes
through to Reaper so that for instance spacebar can invoke play.
This whole thing is a workaround the fact that modifier keys and
ordinary key strokes are treated the same, but should not; mod
keys should be received even when the script is not focused.