Fix Show FX type
The Show FX type feature has been reworked, it now works better.
There are still some issues, since Reaper does not really care
about the difference between VST and VSTi. The main problem is
that the use may have edited the FX names to remove the prefix
(VST: VSTi: JS:) and when getting the FX type from Reaper, we
get 3 for both VST and VSTi. So if the user edited to remove the
prefix of VSTi, we can only add VST as prefix for it. So if the
FX has a prefix in its name, than thi sprefix is used, but if not
then the prefix type given by Reaper determines the prefix.

Some internal changes made, such as changing the numbers for
the various fonts. This is to prepare for showing different FX
types in different fonts and/or colors. Also the fucntion split()
was removed, as it wreaked more havoc than it solved; many FX
have "/" in their names that we do not want to split. Notably
my own "MF/xxx" JS-plugins, and several video processor presets.